refactor: move EndFiniteBurn class to new end_finite_burn.py
