+++ gmatpyplus/command/__init__.py
@@ -1,6 +1,7 @@
 from .begin_finite_burn import BeginFiniteBurn
 from .branch_command import *
 from .command import *
+from .end_finite_burn import EndFiniteBurn
 from .end_target import *
 from .gmat_command import GmatCommand
 from .maneuver import Maneuver
