+++ gmatpyplus/command/command.py
@@ -135,17 +135,6 @@ class BeginMissionSequence(GmatCommand):
 
         self.Initialize()
 
-
-class EndFiniteBurn(GmatCommand):
-    def __init__(self, burn: gp.FiniteBurn | gmat.FiniteBurn, name: str):
-        super().__init__('EndFiniteBurn', name)
-
-        # Assign the user-provided FiniteBurn to this command
-        self.burn = burn
-        self.SetRefObjectName(gmat.FINITE_BURN, self.burn.GetName())
-
-        self.Initialize()
-
 # class PropagateMulti(Propagate):
 #     # TODO: consider making this a nested/inner class of Propagate, so would call Propagate.Multi()
 #     """
